pyproject.toml
[tool.poetry]
name = "social-network-api"
version = "0.1.0"
description = ""
authors = ["a.cheremushkin <a.ch3re359@yandex.ru>"]
readme = "README.md"
packages = [{include = "social_network_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
asyncpg = "^0.27.0"
sqlalchemy = "^1.4.46"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.0"
python-multipart = "^0.0.5"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
jose = "^1.0.0"


[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.18.0"
isort = "^5.11.4"
loguru = "^0.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
